Backend/src/publicador/requirements.txt
annotated-types==0.7.0
paho-mqtt==2.1.0
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
typing_extensions==4.12.2
